+++ Data/LightningCustomPluginTemplate/RESOURCE_NAME_.sln
@@ -1,8 +1,6 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31112.23
-MinimumVisualStudioVersion = 10.0.40219.1
+# Visual Studio Version 17
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "RESOURCE_NAME_", "RESOURCE_NAME_.vcxproj", "{11111111-1111-1111-1111-111111111111}"
 EndProject
 Global
