+++ Data/LightningCustomPluginTemplate/RESOURCE_NAME_.vcxproj
@@ -1,5 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<Project DefaultTargets="Build" ToolsVersion="14.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+<Project DefaultTargets="Build" ToolsVersion="17.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <PreferredToolArchitecture>x64</PreferredToolArchitecture>
+  </PropertyGroup>
   <Import Project="RESOURCE_NAME_Version.props"/>
   <ItemGroup Label="ProjectConfigurations">
     <ProjectConfiguration Include="Debug|x64">
@@ -14,20 +17,23 @@
   <PropertyGroup Label="Globals">
     <ProjectGuid>{11111111-1111-1111-1111-111111111111}</ProjectGuid>
     <RootNamespace>RESOURCE_NAME_</RootNamespace>
+    <Platform>x64</Platform>
+    <ProjectName>RESOURCE_NAME_</ProjectName>
+    <VCProjectUpgraderObjectName>NoUpgrade</VCProjectUpgraderObjectName>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'" Label="Configuration">
     <ConfigurationType>DynamicLibrary</ConfigurationType>
     <UseDebugLibraries>true</UseDebugLibraries>
-    <CharacterSet>NotSet</CharacterSet>
-    <PlatformToolset>v142</PlatformToolset>
+    <CharacterSet>Unicode</CharacterSet>
+    <PlatformToolset>v143</PlatformToolset>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'" Label="Configuration">
     <ConfigurationType>DynamicLibrary</ConfigurationType>
     <UseDebugLibraries>false</UseDebugLibraries>
     <WholeProgramOptimization>true</WholeProgramOptimization>
-    <CharacterSet>NotSet</CharacterSet>
-    <PlatformToolset>v142</PlatformToolset>
+    <CharacterSet>Unicode</CharacterSet>
+    <PlatformToolset>v143</PlatformToolset>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.props" />
   <ImportGroup Label="ExtensionSettings">
@@ -43,32 +49,33 @@
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <TargetName>$(ProjectName)</TargetName>
-    <TargetExt>.$(OS)-$(PlatformShortName)-$(PlasmaVersion)-lightningPlugin</TargetExt>
+    <TargetExt>.Windows-$(PlatformShortName)-$(PlasmaVersion)-lightningPlugin</TargetExt>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <TargetName>$(ProjectName)</TargetName>
-    <TargetExt>.$(OS)-$(PlatformShortName)-$(PlasmaVersion)-lightningPlugin</TargetExt>
+    <TargetExt>.Windows-$(PlatformShortName)-$(PlasmaVersion)-lightninglugin</TargetExt>
   </PropertyGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <ClCompile>
       <WarningLevel>Level3</WarningLevel>
       <Optimization>Disabled</Optimization>
-      <AdditionalIncludeDirectories>$(ProjectDir);Versions\$(OS)-$(PlatformShortName)-$(PlasmaVersion)\;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(ProjectDir);Versions\Windows-$(PlatformShortName)-$(PlasmaVersion)\;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <RuntimeLibrary>MultiThreaded</RuntimeLibrary>
       <AdditionalOptions>/bigobj %(AdditionalOptions)</AdditionalOptions>
       <PrecompiledHeader>Use</PrecompiledHeader>
       <PrecompiledHeaderFile>RESOURCE_NAME_Precompiled.hpp</PrecompiledHeaderFile>
+      <PreprocessorDefinitions>WelderCompilerMsvc;WelderTargetOsWindows;%(PreprocessorDefinitions)</PreprocessorDefinitions>
     </ClCompile>
     <Link>
       <GenerateDebugInformation>true</GenerateDebugInformation>
-      <AdditionalDependencies>Versions\$(OS)-$(PlatformShortName)-$(PlasmaVersion)\PlasmaEditor.lib;%(AdditionalDependencies)</AdditionalDependencies>
+      <AdditionalDependencies>Versions\Windows-$(PlatformShortName)-$(PlasmaVersion)\WelderEditor.lib;%(AdditionalDependencies)</AdditionalDependencies>
     </Link>
     <PostBuildEvent>
       <Command>COPY /Y "$(TargetPath)" "$(ProjectDir)..\$(TargetFileName)"
 PUSHD "$(ProjectDir)..\"
-COPY /B "$(TargetName).lightningPluginSource.data"+
+COPY /B "$(TargetName).LightningPluginSource.data"+
 POPD
 </Command>
       <Message>Copy the dll to the Plasma Project's "Content" directory, and touch the content data file</Message>
@@ -89,17 +96,18 @@ EXIT 0
       <Optimization>MaxSpeed</Optimization>
       <FunctionLevelLinking>true</FunctionLevelLinking>
       <IntrinsicFunctions>true</IntrinsicFunctions>
-      <AdditionalIncludeDirectories>$(ProjectDir);Versions\$(OS)-$(PlatformShortName)-$(PlasmaVersion)\;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(ProjectDir);Versions\Windows-$(PlatformShortName)-$(PlasmaVersion)\;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <RuntimeLibrary>MultiThreaded</RuntimeLibrary>
       <AdditionalOptions>/bigobj %(AdditionalOptions)</AdditionalOptions>
       <PrecompiledHeader>Use</PrecompiledHeader>
       <PrecompiledHeaderFile>RESOURCE_NAME_Precompiled.hpp</PrecompiledHeaderFile>
+      <PreprocessorDefinitions>WelderCompilerMsvc;WelderTargetOsWindows;_WINDLL;%(PreprocessorDefinitions)</PreprocessorDefinitions>
     </ClCompile>
     <Link>
       <GenerateDebugInformation>true</GenerateDebugInformation>
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
-      <AdditionalDependencies>Versions\$(OS)-$(PlatformShortName)-$(PlasmaVersion)\PlasmaEditor.lib;%(AdditionalDependencies)</AdditionalDependencies>
+      <AdditionalDependencies>Versions\Windows-$(PlatformShortName)-$(PlasmaVersion)\WelderEditor.lib;%(AdditionalDependencies)</AdditionalDependencies>
     </Link>
     <PostBuildEvent>
       <Command>COPY /Y "$(TargetPath)" "$(ProjectDir)..\$(TargetFileName)"
@@ -131,19 +139,19 @@ EXIT 0
     <ClInclude Include="RESOURCE_NAME_Precompiled.hpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="Versions\$(OS)-$(PlatformShortName)-$(PlasmaVersion)\PlasmaEngine.cpp" />
+    <ClCompile Include="Versions\Windows-$(PlatformShortName)-$(PlasmaVersion)\PlasmaEngine.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Versions\$(OS)-$(PlatformShortName)-$(PlasmaVersion)\PlasmaEngine.hpp" />
+    <ClInclude Include="Versions\Windows-$(PlatformShortName)-$(PlasmaVersion)\PlasmaEngine.hpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="Versions\$(OS)-$(PlatformShortName)-$(PlasmaVersion)\Core.cpp" />
+    <ClCompile Include="Versions\Windows-$(PlatformShortName)-$(PlasmaVersion)\Core.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Versions\$(OS)-$(PlatformShortName)-$(PlasmaVersion)\Core.hpp" />
+    <ClInclude Include="Versions\Windows-$(PlatformShortName)-$(PlasmaVersion)\Core.hpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Versions\$(OS)-$(PlatformShortName)-$(PlasmaVersion)\Lightning.hpp" />
+    <ClInclude Include="Versions\Windows-$(PlatformShortName)-$(PlasmaVersion)\Lightning.hpp" />
   </ItemGroup>
   <ItemGroup>
     <Natvis Include="PlasmaVisualizers.natvis" />
