+++ Data/LightningCustomPluginTemplate/RESOURCE_NAME_.vcxproj.filters
@@ -1,10 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+<Project ToolsVersion="17.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <ClCompile Include="Versions\$(OS)-$(PlatformShortName)-$(PlasmaVersion)\PlasmaEngine.cpp">
+    <ClCompile Include="Versions\Windows-x64-$(PlasmaVersion)\PlasmaEngine.cpp">
       <Filter>Plasma</Filter>
     </ClCompile>
-    <ClCompile Include="Versions\$(OS)-$(PlatformShortName)-$(PlasmaVersion)\Core.cpp">
+    <ClCompile Include="Versions\Windows-x64-$(PlasmaVersion)\Core.cpp">
       <Filter>Plasma</Filter>
     </ClCompile>
     <ClCompile Include="RESOURCE_NAME_.cpp">
@@ -15,13 +15,13 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Versions\$(OS)-$(PlatformShortName)-$(PlasmaVersion)\PlasmaEngine.hpp">
+    <ClInclude Include="Versions\Windows-x64-$(PlasmaVersion)\PlasmaEngine.hpp">
       <Filter>Plasma</Filter>
     </ClInclude>
-    <ClInclude Include="Versions\$(OS)-$(PlatformShortName)-$(PlasmaVersion)\Core.hpp">
+    <ClInclude Include="Versions\Windows-x64-$(PlasmaVersion)\Core.hpp">
       <Filter>Plasma</Filter>
     </ClInclude>
-    <ClInclude Include="Versions\$(OS)-$(PlatformShortName)-$(PlasmaVersion)\Lightning.hpp">
+    <ClInclude Include="Versions\Windows-x64-$(PlasmaVersion)\Lightning.hpp">
       <Filter>Plasma</Filter>
     </ClInclude>
     <ClInclude Include="RESOURCE_NAME_.hpp">
