+++ Data/LightningCustomPluginTemplate/RESOURCE_NAME_InstallToolsAndRun.bat
@@ -2,5 +2,5 @@
 REM Installs the Visual Studio plugin and then runs our solution
 PUSHD "%~dp0"
 PlasmaLightningPlugins.vsix
-DEL /F /Q PlasmaLightninghPlugins.vsix
+DEL /F /Q PlasmaLightningPlugins.vsix
 "RESOURCE_NAME_.bat"
