+++ Source/Editor/PlasmaEditor/CMakeLists.txt
@@ -65,3 +65,11 @@ target_link_libraries(PlasmaEditor
 target_compile_definitions(PlasmaEditor PUBLIC TRACY_IMPORTS)
 
 plasma_copy_from_linked_libraries(PlasmaEditor)
+
+set_target_properties(PlasmaEditor PROPERTIES ENABLE_EXPORTS TRUE)
+set(platform_plugin_lib_dir ${PLASMA_LIGHTNING_CUSTOM_PLUGIN_SHARED_DIR}${PLASMA_PLATFORM}-${PLASMA_ARCHITECTURE}/)
+add_custom_command(TARGET PlasmaEditor POST_BUILD COMMAND
+    ${CMAKE_COMMAND} -E copy $<TARGET_LINKER_IMPORT_FILE:PlasmaEditor> ${platform_plugin_lib_dir})
+    
+add_custom_command(TARGET PlasmaEditor POST_BUILD COMMAND
+    ${PLASMA_ALL_TO_ONE_TOOL} -Plasma ${PLASMA_CORE_DIR} ${platform_plugin_lib_dir}Lightning.hpp)
