--- Tools/AllToOneCppSource/AllToOneCpp.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31112.23
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AllToOneCpp", "AllToOneCpp\AllToOneCpp.csproj", "{663F9562-9B5E-4FA6-A67D-88193F978297}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LightningToOneCpp", "LightningToOneCpp\LightningToOneCpp.csproj", "{A4952CE1-D6A7-4A99-9211-DECDF8A51AD1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x86 = Debug|x86
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{663F9562-9B5E-4FA6-A67D-88193F978297}.Debug|x86.ActiveCfg = Debug|x86
-		{663F9562-9B5E-4FA6-A67D-88193F978297}.Debug|x86.Build.0 = Debug|x86
-		{663F9562-9B5E-4FA6-A67D-88193F978297}.Release|x86.ActiveCfg = Release|x86
-		{663F9562-9B5E-4FA6-A67D-88193F978297}.Release|x86.Build.0 = Release|x86
-		{A4952CE1-D6A7-4A99-9211-DECDF8A51AD1}.Debug|x86.ActiveCfg = Debug|x86
-		{A4952CE1-D6A7-4A99-9211-DECDF8A51AD1}.Debug|x86.Build.0 = Debug|x86
-		{A4952CE1-D6A7-4A99-9211-DECDF8A51AD1}.Release|x86.ActiveCfg = Release|x86
-		{A4952CE1-D6A7-4A99-9211-DECDF8A51AD1}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9DA60277-9DD3-4C61-ACCF-F5011FB9C042}
-	EndGlobalSection
-EndGlobal
