+++ Utilities/AllToOneCpp/.gitignore
@@ -1,0 +1,3 @@
+.vs/
+bin/
+obj/
