+++ Utilities/AllToOneCpp/AllToOneCpp.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AllToOneCpp", "AllToOneCpp\AllToOneCpp.csproj", "{6114895E-9007-4DDC-9230-561D8B24028D}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LightningToOneCpp", "LightningToOneCpp\LightningToOneCpp.csproj", "{DB6DC07B-BED9-4BDC-9508-2B6812452D9E}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6114895E-9007-4DDC-9230-561D8B24028D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6114895E-9007-4DDC-9230-561D8B24028D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6114895E-9007-4DDC-9230-561D8B24028D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6114895E-9007-4DDC-9230-561D8B24028D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DB6DC07B-BED9-4BDC-9508-2B6812452D9E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DB6DC07B-BED9-4BDC-9508-2B6812452D9E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DB6DC07B-BED9-4BDC-9508-2B6812452D9E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DB6DC07B-BED9-4BDC-9508-2B6812452D9E}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
