+++ Utilities/AllToOneCpp/LightningToOneCpp/LightningToOneCpp.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net7.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\AllToOneCpp\AllToOneCpp.csproj" />
+  </ItemGroup>
+
+</Project>
